[Feature] Add wrap sever service and workflow orchestration
