--- apps/src/main/scala/com/crib/bills/dom6maps/apps/HelloApp.scala
@@ -1,9 +1,0 @@
-package com.crib.bills.dom6maps.apps
-
-object HelloApp extends Greeting with App {
-  println(greeting)
-}
-
-trait Greeting {
-  lazy val greeting: String = "hello"
-}
