--- apps/src/main/scala/com/crib/bills/dom6maps/apps/MapEditorApp.scala
@@ -1,7 +1,0 @@
-package com.crib.bills.dom6maps.apps
-
-import com.crib.bills.dom6maps.model.{Province, ProvinceId}
-
-object MapEditorApp extends App {
-  println(Province(ProvinceId(0), "Example Province"))
-}
