+++ apps/src/main/scala/com/crib/bills/dom6maps/services/mapeditor/PathsConfig.scala
@@ -1,0 +1,8 @@
+package com.crib.bills.dom6maps
+package apps.services.mapeditor
+
+import pureconfig.ConfigReader
+import pureconfig.generic.derivation.default.*
+import java.nio.file.Path as NioPath
+
+final case class PathsConfig(source: NioPath, dest: NioPath) derives ConfigReader
