--- apps/src/test/scala/com/crib/bills/dom6maps/apps/HelloAppSpec.scala
@@ -1,9 +1,0 @@
-package com.crib.bills.dom6maps.apps
-
-import cats.effect.IO
-import weaver.SimpleIOSuite
-
-object HelloAppSpec extends SimpleIOSuite:
-  test("greeting is hello") {
-    IO.pure(expect(HelloApp.greeting == "hello"))
-  }
